Rename to parametrized tests, decouple from system-tests
