+++ .github/workflows/ci.yml
@@ -550,7 +550,7 @@ jobs:
       - name: Display
         run: python scenarios/perfs/process.py
 
-  integration_unit:
+  parametric:
     runs-on: ubuntu-latest
     needs: lint_and_test
     strategy:
@@ -567,8 +567,8 @@ jobs:
       - name: Install
         run: |
           pip install wheel
-          pip install -r requirements.txt
       - name: Run
         run: |
-          cd integration_unit
+          cd parametric
+          pip install -r requirements.txt
           CLIENTS_ENABLED=${{ matrix.client }} ./run.sh
