+++ parametric/apps/dotnet/ApmTestClient.csproj
