+++ parametric/apps/dotnet/appsettings.json
