+++ parametric/apps/golang/generate_protos.sh
