+++ parametric/apps/python/generate_protos.sh
