+++ requirements.txt
@@ -10,6 +10,8 @@ flake8==3.9.2
 python-dateutil==2.8.1
 packaging==20.4
 msgpack==1.0.2
+protobuf==3.16
+grpcio-tools==1.43.0
 
 aiohttp==3.7.4
 yarl==1.4.2
@@ -18,11 +20,3 @@ jsonschema==3.2.0
 rfc3339-validator==0.1.2
 
 matplotlib
-
-
-grpcio-tools==1.47.0
-grpcio==1.47.0
-pytest
-ddtrace
-ddsketch
-numpy
